Removing TxnConfirmationScreenWithData
