+++ src/screens/TransactionConfirmationScreen.js
@@ -1,8 +1,13 @@
+import analytics from '@segment/analytics-react-native';
+import { ethers } from 'ethers';
 import lang from 'i18n-js';
-import { get } from 'lodash';
+import { get, isNil, omit } from 'lodash';
 import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
-import { Animated } from 'react-native';
+import { Alert, Animated, InteractionManager, Vibration } from 'react-native';
+import { isEmulatorSync } from 'react-native-device-info';
+import { withNavigationFocus } from 'react-navigation';
+import { compose } from 'recompact';
 import styled from 'styled-components';
 import { Button, HoldToAuthorizeButton } from '../components/buttons';
 import { RequestVendorLogoIcon } from '../components/coin-icon';
@@ -13,11 +18,26 @@ import {
   MessageSigningSection,
   TransactionConfirmationSection,
 } from '../components/transaction';
+import { estimateGas, getTransactionCount, toHex } from '../handlers/web3';
+import { withGas, withTransactionConfirmationScreen } from '../hoc';
+import {
+  sendTransaction,
+  signMessage,
+  signPersonalMessage,
+  signTransaction,
+  signTypedDataMessage,
+} from '../model/wallet';
 import { colors, position } from '../styles';
+import { gasUtils, logger } from '../utils';
 import {
   isMessageDisplayType,
+  isSignFirstParamType,
+  isSignSecondParamType,
   isTransactionDisplayType,
+  PERSONAL_SIGN,
   SEND_TRANSACTION,
+  SIGN,
+  SIGN_TYPED_DATA,
 } from '../utils/signingMethods';
 
 const CancelButtonContainer = styled.View`
@@ -42,14 +62,17 @@ const TransactionType = styled(Text).attrs({ size: 'h5' })`
   margin-top: 6;
 `;
 
-export default class TransactionConfirmationScreen extends PureComponent {
+class TransactionConfirmationScreen extends PureComponent {
   static propTypes = {
-    dappName: PropTypes.string,
-    imageUrl: PropTypes.string,
-    method: PropTypes.string,
-    onCancel: PropTypes.func,
-    onConfirm: PropTypes.func,
-    request: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+    dataAddNewTransaction: PropTypes.func,
+    gasPrices: PropTypes.object,
+    gasPricesStartPolling: PropTypes.func,
+    gasPricesStopPolling: PropTypes.func,
+    navigation: PropTypes.any,
+    removeRequest: PropTypes.func,
+    transactionCountNonce: PropTypes.number,
+    updateTransactionCountNonce: PropTypes.func,
+    walletConnectSendStatus: PropTypes.func,
   };
 
   state = {
@@ -57,9 +80,198 @@ export default class TransactionConfirmationScreen extends PureComponent {
     sendLongPressProgress: new Animated.Value(0),
   };
 
+  componentDidMount() {
+    const { navigation } = this.props;
+    const openAutomatically = get(navigation, 'state.params.openAutomatically');
+    if (openAutomatically && !isEmulatorSync()) {
+      Vibration.vibrate();
+    }
+
+    InteractionManager.runAfterInteractions(() => {
+      this.props.gasPricesStartPolling();
+    });
+  }
+
   componentWillUnmount() {
     this.state.sendLongPressProgress.stopAnimation();
   }
+
+  handleConfirmTransaction = async () => {
+    const {
+      callback,
+      transactionDetails: {
+        dappName,
+        displayDetails,
+        payload: { method, params },
+        peerId,
+        requestId,
+      },
+    } = this.props.navigation.state.params;
+
+    const sendInsteadOfSign = method === SEND_TRANSACTION;
+    const txPayload = get(params, '[0]');
+    let { gasLimit, gasPrice } = txPayload;
+
+    if (isNil(gasPrice)) {
+      const { gasPrices } = this.props;
+      const rawGasPrice = get(gasPrices, `${gasUtils.NORMAL}.value.amount`);
+      if (rawGasPrice) {
+        gasPrice = toHex(rawGasPrice);
+      }
+    }
+
+    if (isNil(gasLimit)) {
+      try {
+        const rawGasLimit = await estimateGas(txPayload);
+        gasLimit = toHex(rawGasLimit);
+      } catch (error) {
+        logger.log('error estimating gas', error);
+      }
+    }
+
+    const web3TxnCount = await getTransactionCount(txPayload.from);
+    const maxTxnCount = Math.max(
+      this.props.transactionCountNonce,
+      web3TxnCount
+    );
+    const nonce = ethers.utils.hexlify(maxTxnCount);
+    let txPayloadLatestNonce = { ...txPayload, gasLimit, gasPrice, nonce };
+    txPayloadLatestNonce = omit(txPayloadLatestNonce, 'from');
+    let result = null;
+    if (sendInsteadOfSign) {
+      result = await sendTransaction({
+        transaction: txPayloadLatestNonce,
+      });
+    } else {
+      result = await signTransaction({
+        transaction: txPayloadLatestNonce,
+      });
+    }
+
+    if (result) {
+      if (callback) {
+        callback({ result });
+      }
+      if (sendInsteadOfSign) {
+        this.props.updateTransactionCountNonce(maxTxnCount + 1);
+        const txDetails = {
+          amount: get(displayDetails, 'request.value'),
+          asset: get(displayDetails, 'request.asset'),
+          dappName,
+          from: get(displayDetails, 'request.from'),
+          gasLimit,
+          gasPrice,
+          hash: result,
+          nonce,
+          to: get(displayDetails, 'request.to'),
+        };
+        this.props.dataAddNewTransaction(txDetails);
+      }
+      analytics.track('Approved WalletConnect transaction request');
+      if (requestId) {
+        this.props.removeRequest(requestId);
+        await this.props.walletConnectSendStatus(peerId, requestId, result);
+      }
+      this.closeScreen();
+    } else {
+      await this.handleCancelRequest();
+    }
+  };
+
+  handleSignMessage = async () => {
+    const {
+      callback,
+      transactionDetails: {
+        payload: { method, params },
+        peerId,
+        requestId,
+      },
+    } = this.props.navigation.state.params;
+    let message = null;
+    let flatFormatSignature = null;
+    if (isSignFirstParamType(method)) {
+      message = get(params, '[0]');
+    } else if (isSignSecondParamType(method)) {
+      message = get(params, '[1]');
+    }
+
+    switch (method) {
+      case SIGN:
+        flatFormatSignature = await signMessage(message);
+        break;
+      case PERSONAL_SIGN:
+        flatFormatSignature = await signPersonalMessage(message);
+        break;
+      case SIGN_TYPED_DATA:
+        flatFormatSignature = await signTypedDataMessage(message, method);
+        break;
+      default:
+        break;
+    }
+
+    if (flatFormatSignature) {
+      analytics.track('Approved WalletConnect signature request');
+      if (requestId) {
+        this.props.removeRequest(requestId);
+        await this.props.walletConnectSendStatus(
+          peerId,
+          requestId,
+          flatFormatSignature
+        );
+      }
+      if (callback) {
+        callback({ sig: flatFormatSignature });
+      }
+      this.closeScreen();
+    } else {
+      await this.handleCancelRequest();
+    }
+  };
+
+  onConfirm = async () => {
+    const {
+      transactionDetails: {
+        payload: { method },
+      },
+    } = this.props.navigation.state.params;
+    if (isMessageDisplayType(method)) {
+      return this.handleSignMessage();
+    }
+    return this.handleConfirmTransaction();
+  };
+
+  closeScreen = () => {
+    this.props.navigation.popToTop();
+    this.props.gasPricesStopPolling();
+  };
+
+  onCancel = async () => {
+    try {
+      this.closeScreen();
+      const {
+        callback,
+        transactionDetails: {
+          payload: { method },
+          peerId,
+          requestId,
+        },
+      } = this.props.navigation.state.params;
+      if (callback) {
+        callback({ error: 'User cancelled the request' });
+      }
+      if (requestId) {
+        await this.props.walletConnectSendStatus(peerId, requestId, null);
+        this.props.removeRequest(requestId);
+      }
+      const rejectionType =
+        method === SEND_TRANSACTION ? 'transaction' : 'signature';
+      analytics.track(`Rejected WalletConnect ${rejectionType} request`);
+    } catch (error) {
+      logger.log('error while handling cancel request', error);
+      this.closeScreen();
+      Alert.alert(lang.t('wallet.transaction.alert.cancelled_transaction'));
+    }
+  };
 
   onPressSend = () => {
     const { sendLongPressProgress } = this.state;
@@ -80,7 +292,6 @@ export default class TransactionConfirmationScreen extends PureComponent {
   };
 
   onLongPressSend = async () => {
-    const { onConfirm } = this.props;
     const { sendLongPressProgress } = this.state;
 
     this.setState({ isAuthorizing: true });
@@ -90,7 +301,7 @@ export default class TransactionConfirmationScreen extends PureComponent {
     }).start();
 
     try {
-      await onConfirm();
+      await this.onConfirm();
       this.setState({ isAuthorizing: false });
     } catch (error) {
       this.setState({ isAuthorizing: false });
@@ -98,7 +309,11 @@ export default class TransactionConfirmationScreen extends PureComponent {
   };
 
   renderSendButton = () => {
-    const { method } = this.props;
+    const {
+      transactionDetails: {
+        payload: { method },
+      },
+    } = this.props.navigation.state.params;
     const { isAuthorizing } = this.state;
     const label = `Hold to ${method === SEND_TRANSACTION ? 'Send' : 'Sign'}`;
 
@@ -112,14 +327,23 @@ export default class TransactionConfirmationScreen extends PureComponent {
   };
 
   requestHeader = () => {
-    const { method } = this.props;
+    const {
+      transactionDetails: {
+        payload: { method },
+      },
+    } = this.props.navigation.state.params;
     return isMessageDisplayType(method)
       ? lang.t('wallet.message_signing.request')
       : lang.t('wallet.transaction.request');
   };
 
   renderTransactionSection = () => {
-    const { request, method } = this.props;
+    const {
+      transactionDetails: {
+        displayDetails: { request },
+        payload: { method },
+      },
+    } = this.props.navigation.state.params;
 
     if (isMessageDisplayType(method)) {
       return (
@@ -158,38 +382,50 @@ export default class TransactionConfirmationScreen extends PureComponent {
     );
   };
 
-  render = () => (
-    <Container>
-      <Masthead>
-        <RequestVendorLogoIcon
-          backgroundColor="transparent"
-          dappName={this.props.dappName}
-          imageUrl={this.props.imageUrl}
-          size={60}
-          style={{ marginBottom: 24 }}
-        />
-        <Text
-          color="white"
-          letterSpacing="roundedMedium"
-          size="h4"
-          weight="semibold"
-        >
-          {this.props.dappName}
-        </Text>
-        <TransactionType>{this.requestHeader()}</TransactionType>
-        <CancelButtonContainer>
-          <Button
-            backgroundColor={colors.alpha(colors.grey, 0.4)}
-            onPress={this.props.onCancel}
-            showShadow={false}
-            size="small"
-            textProps={{ color: colors.black, size: 'lmedium' }}
+  render = () => {
+    const {
+      transactionDetails: { dappName, imageUrl },
+    } = this.props.navigation.state.params;
+
+    return (
+      <Container>
+        <Masthead>
+          <RequestVendorLogoIcon
+            backgroundColor="transparent"
+            dappName={dappName || ''}
+            imageUrl={imageUrl || ''}
+            size={60}
+            style={{ marginBottom: 24 }}
+          />
+          <Text
+            color="white"
+            letterSpacing="roundedMedium"
+            size="h4"
+            weight="semibold"
           >
-            {lang.t('wallet.action.reject')}
-          </Button>
-        </CancelButtonContainer>
-      </Masthead>
-      {this.renderTransactionSection()}
-    </Container>
-  );
+            {dappName}
+          </Text>
+          <TransactionType>{this.requestHeader()}</TransactionType>
+          <CancelButtonContainer>
+            <Button
+              backgroundColor={colors.alpha(colors.grey, 0.4)}
+              onPress={this.onCancel}
+              showShadow={false}
+              size="small"
+              textProps={{ color: colors.black, size: 'lmedium' }}
+            >
+              {lang.t('wallet.action.reject')}
+            </Button>
+          </CancelButtonContainer>
+        </Masthead>
+        {this.renderTransactionSection()}
+      </Container>
+    );
+  };
 }
+
+export default compose(
+  withGas,
+  withNavigationFocus,
+  withTransactionConfirmationScreen
+)(TransactionConfirmationScreen);
